Split solution into multiple files.
