+++ node-notes-cli/read-notes.js
@@ -1,0 +1,9 @@
+const data = require('./data.json');
+
+const readNotes = () => {
+  for (const key in data.notes) {
+    console.log(key + ': ' + data.notes[key]);
+  }
+};
+
+module.exports = readNotes;
